BigInt

+++ apps/indicator/migrations/0006_auto_20171207_1348.py
@@ -15,6 +15,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='priceresampled',
             name='closing_price',
-            field=models.IntegerField(null=True),
+            field=models.BigIntegerField(null=True),
         ),
     ]
